Added ability to edit a transaction
